modified ioutils and add for modularized code
